imap: move IdleHandle from session.rs to idle.rs
